+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `meta::` queries are no deprecated and can only be used in quirks mode
 
+### Fixed
+
+- Quirks mode: add additional identity joins in the order as the nodes are defined in the query
+
 ## [0.19.4] - 2019-05-10
 
 ### Changed